+++ src/Inane/Http/FileServer.php
@@ -16,14 +16,18 @@
 namespace Inane\Http;
 
 use Inane\File\FileInfo;
+use Inane\Observer\InaneSubject;
+use Inane\Observer\InaneObserver;
 
 /**
  * Serve file over http with resume support
  * 
  * @package Inane\Http\FileServer
- * @version 0.6.0
+ * @version 0.7.0
  */
-class FileServer {
+class FileServer extends InaneSubject {
+	private $observers = [];
+	
 	/**
 	 * File Information
 	 * 
@@ -44,6 +48,57 @@ class FileServer {
 	 * @var bool
 	 */
 	protected $_forceDownload = true;
+	
+	/**
+	 * Limit speed of transfer (0 = unlimited)
+	 *
+	 * @var int
+	 */
+	protected $_bandwidth = 0;
+	
+	/**
+	 * Sleep time to limit bandwidth
+	 *
+	 * @var int
+	 */
+	protected $_sleep = 0;
+	
+	/**
+	 * File size served
+	 *
+	 * @var int
+	 */
+	protected $_progress = 0;
+	protected $_percent = 0;
+
+	/**
+	 * Progress of download
+	 *
+	 * @return [];
+	 */
+	public function getProgress() {
+		return [
+			'filename' => $this->_file->getFilename(),
+			'progress' => $this->_progress,
+			'total' => $this->_file->getSize()];
+	}
+
+	/**
+	 * @param int $progress
+	 * @return \Inane\Http\FileServer
+	 */
+	protected function addProgress($progress) {
+		$this->_progress += $progress;
+		if ($this->_progress > $this->_file->getSize())
+			$this->_progress = $this->_file->getSize();
+		
+		$percent = round($this->_progress / $this->_file->getSize() * 100, 0);
+		if ($percent != $this->_percent) {
+			$this->notify();
+			$this->_percent = $percent;
+		}
+		return $this;
+	}
 
 	/**
 	 * Prepare a file for serving
@@ -58,7 +113,45 @@ class FileServer {
 		}
 		$this->_file = $file;
 	}
-	
+
+	/**
+	 * {@inheritDoc}
+	 * @see \Inane\Observer\InaneSubject::attach()
+	 */
+	public function attach(InaneObserver $observer_in) {
+		//could also use array_push($this->observers, $observer_in);
+		$this->observers[] = $observer_in;
+		
+		return $this;
+	}
+
+	/**
+	 * {@inheritDoc}
+	 * @see \Inane\Observer\InaneSubject::detach()
+	 */
+	public function detach(InaneObserver $observer_in) {
+		//$key = array_search($observer_in, $this->observers);
+		foreach ( $this->observers as $okey => $oval ) {
+			if ($oval == $observer_in) {
+				unset($this->observers[$okey]);
+			}
+		}
+		
+		return $this;
+	}
+
+	/**
+	 * {@inheritDoc}
+	 * @see \Inane\Observer\InaneSubject::notify()
+	 */
+	public function notify() {
+		foreach ( $this->observers as $obs ) {
+			$obs->update($this);
+		}
+		
+		return $this;
+	}
+
 	/**
 	 * Filename of download
 	 * 
@@ -68,9 +161,9 @@ class FileServer {
 		if (isset($this->_name))
 			return $this->_name;
 		
-		return $this->_file->getFilename();	
+		return $this->_file->getFilename();
 	}
-	
+
 	/**
 	 * Set a different name for download
 	 *   or null for realname
@@ -88,7 +181,30 @@ class FileServer {
 		}
 		return $this;
 	}
-	
+
+	/**
+	 * Fets download limit
+	 * 
+	 * @return the int
+	 */
+	public function getBandwidth() {
+		return $this->_bandwidth;
+	}
+
+	/**
+	 * Sets download limit 0 = unlimited
+	 * 
+	 * @param  $_bandwidth
+	 * @return \Inane\Http\FileServer
+	 */
+	public function setBandwidth($_bandwidth) {
+		$this->_sleep = $this->_bandwidth = $_bandwidth;
+		if ($this->_bandwidth > 0)
+			$this->_sleep = (8 / $this->_bandwidth) * 1e6;
+		
+		return $this;
+	}
+
 	/**
 	 * Force files to download and not open in browser
 	 *
@@ -98,10 +214,23 @@ class FileServer {
 	public function forceDownload($state = null) {
 		if ($state === null)
 			return $this->_forceDownload;
-	
+		
 		$this->_forceDownload = $state;
 		
 		return $this;
+	}
+
+	protected function sendHeaders(\Zend\Http\Headers $headers, $status) {
+		$headerArray = explode("\n", $headers->toString());
+		if ($status == 206)
+			header("HTTP/1.1 206 Patial Content");
+		else 
+			header("HTTP/1.1 200 OK");
+		foreach ( $headerArray as $headerLine ) {
+			if (strlen($headerLine) > 0) {
+				header($headerLine);
+			}
+		}
 	}
 
 	/**
@@ -137,6 +266,7 @@ class FileServer {
 			$byte_from = (int) $ranges[0];
 			$byte_to = (int) ($ranges[1] == '' ? $fileSize - 1 : $ranges[1]);
 			$download_size = $byte_to - $byte_from + 1; // the download length
+			
 
 			$download_range = 'bytes ' . $byte_from . "-" . $byte_to . "/" . $fileSize; // the download range
 			$headers->addHeader(new \Zend\Http\Header\ContentRange($download_range));
@@ -164,6 +294,24 @@ class FileServer {
 		// send the file content
 		$fp = fopen($this->_file->getPathname(), "r"); // open the file
 		fseek($fp, $byte_from); // seek to start of missing part
+		$this->_progress = $byte_from;
+		
+		if ($this->_bandwidth > 0) {
+			$this->sendHeaders($headers, $response->getStatusCode());
+			
+			while ( ! feof($fp) ) { // start buffered download
+				set_time_limit(0); // reset time limit for big files (has no effect if php is executed in safe mode)
+				print(fread($fp, 1024 * 8)); // send 8ko
+				flush();
+				$this->addProgress(1024 * 8);
+				usleep($this->_sleep); // sleep (for speed limitation)
+			}
+			$this->addProgress(1);
+			fclose($fp); // close the file
+			
+			exit();
+		}
+		
 		$countent = fread($fp, $download_size);
 		
 		fclose($fp); // close the file